add slice entity and test

+++ services/promotion_expander/src/main/java/com/foodapp/promotion_expander/domain/model/EffectRule.java
@@ -13,5 +13,5 @@ import java.math.BigDecimal;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class EffectRule {
     private String type; // e.g., "PERCENTAGE_OFF_ORDER"
-    private BigDecimal value;
+    private double value;
 }
