+++ services/promotion_service/src/main/java/com/foodapp/promotion_service/domain/model/PromotionRules.java
@@ -57,7 +57,7 @@ public class PromotionRules {
     @AllArgsConstructor
     public static class Effect {
         private String type; // e.g., "PERCENTAGE_OFF_ORDER"
-        private BigDecimal value;
+        private double value;
     }
 
     @Data
